+++ packages/apollo/src/commands/service/push.ts
@@ -33,7 +33,7 @@ export default class ServicePush extends ProjectCommand {
           }
 
           const schema = await project.resolveSchema({ tag: flags.tag });
-          gitContext = await gitInfo();
+          gitContext = await gitInfo(this.log);
 
           const { tag, code } = await project.engine.uploadSchema({
             id: config.name,